refactor(team): extract canonical state-root resolver
